+++ HISTORY.md
@@ -55,6 +55,7 @@
 
 ### 2023-06-02
 * blocks, symbols: made collapse list icon red
+* blocks: tweaked collapse labels
 
 ### 2023-06-01
 * blocks: turned on "explicit input lists" for variadic slots